Relocate admin tool menus, Add chat tool menu
Relocates all admin menus to a single admin drop down button
Removes admin tools from chat header
Refactors admin dialogs to separate files
Creates chat tool menu
Creates structure for chat tool dialogs
Removes old files to simplify naming
Better categorized dialogs for admin and chat tools
Adds link to dashboard
